Progress of Wavemeter ToolChain: -New gui file with subscript _2 -generating main window without QT Designer / .ui file -Implemented InteractiveCurvesWidget such that the current laser position is visible as marker with the move_marker_selection method -Implemented an axis toggle option between nm and Hz
The logic is changed in such a fashion that it needs the time series reader application to be running with only one digital channel. For this an additional signal in the time_series_reader_logic.py is added which emits only the most recent data points of the time series instreamer and is received in the wavemeter_histogram_logic.py.